Hid the admin panel from non-admin users. Added Topics. Added the Tests page. Fixed a bug that caused a crash when a user, whose account got deleted while they were online, tried to access any of the site's pages. Small improvements.

+++ routes.js
@@ -13,17 +13,21 @@ module.exports = function(app) {
                 if (results.length === 0) {
                     req.session.destroy((err) => {
                         res.clearCookie('sessionID');
+                        return res.redirect('/');
                     });
                 }
                 else {
                     // Workaround for session disappearing while request is processing
                     if (req.session && req.session.user)
                         req.session.user.verified = results[0].verified;
+                    
+                    next();
                 }
             });
         }
-
-        next();
+        else {
+            next();
+        }
     });
 
     /* -------- Routing -------- */
@@ -36,7 +40,7 @@ module.exports = function(app) {
         else if (req.query.action === 'deleted_account')
             messages.push('Ваш аккаунт успешно удалён!');
 
-        if (req.session.user)
+        if (req.session && req.session.user)
         {
             if (!req.session.user.verified)
                 messages.push('Ожидайте подтверждения Вашего аккаунта Администратором. Некоторые возможности отключены.');
@@ -102,6 +106,8 @@ module.exports = function(app) {
 
         const admin = (req.session.user.type === 3 && req.session.user.verified) ? true : false;
 
+        var messages = [];
+
         if (req.query.error) {
             if (req.query.error === 'no_user')
                 messages.push('Ошибка: Пользователь не найден.');
@@ -114,8 +120,6 @@ module.exports = function(app) {
             messages.push('Пользователь ' + req.query.deleted + ' успешно удалён.');
         else if (req.query.verified)
             messages.push('Пользователь ' + req.query.verified + ' успешно подтверждён.');
-
-        var messages = [];
 
         const elements_per_page = 10;
         
@@ -150,11 +154,214 @@ module.exports = function(app) {
         });
     });
 
+    app.get('/tests', (req, res) => {
+        if (!req.session.user)
+            return res.redirect('/error');
+
+        const teacher = (req.session.user.type >= 2 && req.session.user.verified) ? true : false;
+
+        var messages = [];
+
+        if (req.query.error) {
+            if (req.query.error === 'no_test')
+                messages.push('Ошибка: Тема не найдена.');
+        }
+        else if (req.query.deleted)
+            messages.push('Тест "' + decodeURIComponent(req.query.deleted) + '" успешно удалён.');
+        else if (req.query.success)
+            messages.push('Тест успешно создан.');
+
+        const elements_per_page = 10;
+        
+        var page = 1;
+        if (req.query.page) {
+            page = parseInt(req.query.page, 10);
+
+            if (page < 1) {
+                messages.push("Страница с введённым номером не существует.");
+                page = 1;
+            }
+        }
+        
+        var pages_num = 1;
+        connection.query('SELECT COUNT(*) FROM tests', (err, results, fields) => {
+            if (err) {
+                console.log('An error has occured on /tests. ' + err.code + ': ' + err.sqlMessage);
+                return res.redirect('/error');
+            }
+
+            pages_num = results[0]['COUNT(*)'] !== 0 ? Math.ceil(results[0]['COUNT(*)'] / elements_per_page) : 0;
+        });
+
+        var offset = (page - 1) * elements_per_page;
+        connection.query('SELECT tests.id, tests.headline, topics.title, topics.subject, users.id AS author_id, users.name, tests.created + INTERVAL 3 HOUR AS created FROM tests LEFT JOIN topics ON tests.topic = topics.id LEFT JOIN users ON tests.author = users.id ORDER BY tests.id DESC LIMIT ? OFFSET ?', [elements_per_page, offset], (err, results, fields) => {
+            if (err) {
+                console.log('An error has occured on /topics. ' + err.code + ': ' + err.sqlMessage);
+                return res.redirect('/error');
+            }
+
+            return res.render('tests', { messages: messages, user: req.session.user, page: page, pages_num: pages_num, elements: results, teacher: teacher });
+        });
+    });
+
+    app.get('/test', (req, res) => {
+        res.redirect('/tests');
+    });
+
+    app.post('/delete_test/:id', (req, res) => {
+        if (!req.session.user || req.session.user.type < 2 || !req.session.user.verified)
+            return res.redirect('/error');
+
+        connection.query('SELECT headline FROM tests WHERE id = ?', [req.params.id], (err, results, fields) => {
+            if (err) {
+                console.log('An error has occured on /delete_test. ' + err.code + ': ' + err.sqlMessage);
+                return res.redirect('/error');
+            }
+
+            if (results.length === 0)
+                return res.redirect('/tests?error=no_test');
+
+            var headline = results[0].headline;
+
+            connection.query('DELETE FROM tests WHERE id = ?', [req.params.id], (err, results, fields) => {
+                if (err) {
+                    console.log('An error has occured on /delete_test. ' + err.code + ': ' + err.sqlMessage);
+                    return res.redirect('/error');
+                }
+    
+                return res.redirect('/tests?deleted=' + encodeURIComponent(headline));
+            });
+        });
+    });
+
+    app.get('/topics', (req, res) => {
+        if (!req.session.user)
+            return res.redirect('/error');
+
+        const teacher = (req.session.user.type >= 2 && req.session.user.verified) ? true : false;
+
+        var messages = [];
+
+        if (req.query.error) {
+            if (req.query.error === 'no_topic')
+                messages.push('Ошибка: Тема не найдена.');
+        }
+        else if (req.query.deleted)
+            messages.push('Тема "' + decodeURIComponent(req.query.deleted) + '" успешно удалена.');
+        else if (req.query.success)
+            messages.push('Тема успешно создана.');
+
+        const elements_per_page = 10;
+        
+        var page = 1;
+        if (req.query.page) {
+            page = parseInt(req.query.page, 10);
+
+            if (page < 1) {
+                messages.push("Страница с введённым номером не существует.");
+                page = 1;
+            }
+        }
+        
+        var pages_num = 1;
+        connection.query('SELECT COUNT(*) FROM topics', (err, results, fields) => {
+            if (err) {
+                console.log('An error has occured on /topics. ' + err.code + ': ' + err.sqlMessage);
+                return res.redirect('/error');
+            }
+
+            pages_num = results[0]['COUNT(*)'] !== 0 ? Math.ceil(results[0]['COUNT(*)'] / elements_per_page) : 0;
+        });
+
+        var offset = (page - 1) * elements_per_page;
+        connection.query('SELECT id, title, subject, semester FROM topics LIMIT ? OFFSET ?', [elements_per_page, offset], (err, results, fields) => {
+            if (err) {
+                console.log('An error has occured on /topics. ' + err.code + ': ' + err.sqlMessage);
+                return res.redirect('/error');
+            }
+
+            return res.render('topics', { messages: messages, user: req.session.user, page: page, pages_num: pages_num, elements: results, teacher: teacher });
+        });
+    });
+
+    app.post('/delete_topic/:id', (req, res) => {
+        if (!req.session.user || req.session.user.type < 2 || !req.session.user.verified)
+            return res.redirect('/error');
+
+        connection.query('SELECT title FROM topics WHERE id = ?', [req.params.id], (err, results, fields) => {
+            if (err) {
+                console.log('An error has occured on /delete_topic. ' + err.code + ': ' + err.sqlMessage);
+                return res.redirect('/error');
+            }
+
+            if (results.length === 0)
+                return res.redirect('/topics?error=no_topic');
+
+            var title = results[0].title;
+
+            connection.query('DELETE FROM topics WHERE id = ?; DELETE FROM tests WHERE topic = ?', [req.params.id, req.params.id], (err, results, fields) => {
+                if (err) {
+                    console.log('An error has occured on /delete_topic. ' + err.code + ': ' + err.sqlMessage);
+                    return res.redirect('/error');
+                }
+    
+                return res.redirect('/topics?deleted=' + encodeURIComponent(title));
+            });
+        });
+    });
+
+    app.get('/create_topic', (req, res) => {
+        if (!req.session.user || req.session.user.type < 2 || !req.session.user.verified)
+            return res.redirect('/error');
+
+        return res.render('create_topic', { user: req.session.user });
+    });
+
+    app.post('/create_topic', (req, res) => {
+        if (!req.session.user || req.session.user.type < 2 || !req.session.user.verified)
+            return res.redirect('/error');
+
+        var title = req.body.inputTitle;
+        var subject = req.body.inputSubject;
+        var semester = req.body.inputSemester;
+
+        if (!title)
+            return res.render('create_topic', { messages: ["Введите название темы."], user: req.session.user });
+        title = escapeHTML(title);
+        if (title.length < 3)
+            return res.render('create_topic', { messages: ["Название темы не может быть короче 3 символов."], user: req.session.user });
+        
+        if (!subject)
+            return res.render('create_topic', { messages: ["Введите название предмета."], user: req.session.user });
+        if (!(/^[а-яА-Я\s]*$/.test(subject)))
+            return res.render('create_topic', { messages: ["В названии предмета допустима только кириллица."], user: req.session.user });
+        if (subject.length < 2)
+            return res.render('create_topic', { messages: ["Название предмета не может быть короче 2 символов."], user: req.session.user });
+        
+        if (!semester)
+            return res.render('create_topic', { messages: ["Выберите необходимый уровень подготовки."], user: req.session.user });
+        semester = parseInt(semester, 10);
+        if (semester === -1)
+            return res.render('create_topic', { messages: ["Выберите необходимый уровень подготовки."], user: req.session.user });
+        if (semester < 0)
+            return res.render('create_topic', { messages: ["Неверный уровень подготовки."], user: req.session.user });
+
+        connection.query('INSERT INTO topics (title, subject, semester) VALUES (?, ?, ?)', 
+                        [title, subject, semester], (err, results, fields) => {
+            if (err) {
+                console.log('An error has occured on /create_topic. ' + err.code + ': ' + err.sqlMessage);
+                return res.redirect('/error');
+            }
+
+            return res.redirect('/topics?success=true');
+        });
+    });
+
     app.post('/delete_account', (req, res) => {
         if (!req.session.user)
             return res.redirect('/error');
 
-        connection.query('DELETE FROM users WHERE id = ?', [req.session.user.id], (err, results, fields) => {
+        connection.query('DELETE FROM users WHERE id = ?; DELETE FROM tests WHERE author = ?', [req.session.user.id, req.session.user.id], (err, results, fields) => {
             if (err) {
                 console.log('An error has occured on /delete_account. ' + err.code + ': ' + err.sqlMessage);
                 return res.redirect('/error');
@@ -184,7 +391,7 @@ module.exports = function(app) {
 
             var login = results[0].login;
 
-            connection.query('DELETE FROM users WHERE id = ?', [req.params.id], (err, results, fields) => {
+            connection.query('DELETE FROM users WHERE id = ?; DELETE FROM tests WHERE author = ?', [req.params.id, req.params.id], (err, results, fields) => {
                 if (err) {
                     console.log('An error has occured on /admin/delete_account. ' + err.code + ': ' + err.sqlMessage);
                     return res.redirect('/error');
@@ -300,7 +507,7 @@ module.exports = function(app) {
             return res.render('register', { messages: ["Введите логин и пароль."] });
         if (login.length < 4)
             return res.render('register', { messages: ["Логин не может быть короче 4 символов."] });
-        if (!(/^[a-zA-z]+$/.test(login)))
+        if (!(/^[a-zA-Z]+$/.test(login)))
             return res.render('register', { messages: ["Логин не должен содержать пробелов. Допустимы только латинские буквы."] });
         if (password.length < 4)
             return res.render('register', { messages: ["Пароль не может быть короче 4 символов."] });
@@ -325,6 +532,7 @@ module.exports = function(app) {
         
         if (!position)
             return res.render('register', { messages: ["Введите ваш класс/должность."] });
+        position = escapeHTML(position);
 
         var hash = bcrypt.hashSync(password, bcrypt.genSaltSync(10));
 
@@ -365,3 +573,15 @@ module.exports = function(app) {
         return res.render('error');
     });
 };
+
+function escapeHTML(text) {
+    var map = {
+        '&': '&amp;',
+        '<': '&lt;',
+        '>': '&gt;',
+        '"': '&quot;',
+        "'": '&#039;'
+    };
+
+    return text.replace(/[&<>"']/g, (m) => { return map[m]; });
+}
